Adds SDK Version to user-agent (#49)
Main purpose:

Adds a user agent to all HTTP requests. Example: Example---Local/1.0.1 (iPhone; iOS Version 16.2 (Build 20C52)) attentive-ios-sdk/0.3.2

Additional updates:

Shows the domain in the Example app UI now
Updates the name of the local attentive-ios-sdk framework so there are no collisions with the Pod framework name. The issue is they both want to use the name attentive-ios-sdk, but only one can. I changed the local framework name to attentive-ios-sdk-framework since it's only being used by our Example apps.
Now that the name collision is fixed, I changed back to use_framework for the Pods. This creates a dynamic framework instead of a static framework, which is likely what our customers will use. The deleted files in this PR are due to the removal of the static framework.
Updated the SDK version that the Example apps rely on